Update Repositories to use Integer in ID

+++ scheduler-server/src/main/java/xyz/webspaghetti/schedulerserver/repository/TaskRepository.java
@@ -7,8 +7,8 @@ import xyz.webspaghetti.schedulerserver.entity.Task;
 
 import java.util.List;
 
-public interface TaskRepository extends JpaRepository<Task, Long> {
+public interface TaskRepository extends JpaRepository<Task, Integer> {
 
     @Query("SELECT t FROM Task t JOIN t.users u WHERE u.id = :user_id AND t.team.id = :team_id")
-    List<Task> findTasksInTeamByUser(@Param("user_id") long userId, @Param("team_id") long teamId);
+    List<Task> findTasksInTeamByUser(@Param("user_id") Integer userId, @Param("team_id") Integer teamId);
 }
